Fix "Read beyond the end of the OBU" errors when reading from large buffers.
  - `ReadAndValidatePayload` compares the amount of bytes read to determine how many "footer" bytes are present in the OBU.
  - The comparison uses the relative offsets from the buffer in bits.
  - Because `expected_final_position` was stored as an int, it would overflow when approaching buffers larger INT_MAX bits (2^31 -1 bits ~= 268 megabytes).
  - The overflow caused the footer logic to fall through `else` branch at the end of `ReadAndValidatePayload` and result in an error.

PiperOrigin-RevId: 676900556

+++ iamf/obu/obu_base.cc
@@ -69,7 +69,7 @@ absl::Status ObuBase::ValidateAndWriteObu(WriteBitBuffer& final_wb) const {
 absl::Status ObuBase::ReadAndValidatePayload(int64_t payload_size_bytes,
                                              ReadBitBuffer& rb) {
   // TODO(b/359588455): Use `ReadBitBuffer::Seek` and `Tell`.
-  const int expected_final_position =
+  const int64_t expected_final_position =
       (rb.source_bit_offset() - (rb.buffer_size() - rb.buffer_bit_offset())) +
       (payload_size_bytes * 8);
